# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Advent of Code Bingo API in Go using Gin framework.",
        "title": "Advent of Code Bingo API",
        "contact": {
            "name": "Jakob Maležič",
            "url": "https://github.com/Blarc"
        },
        "license": {
            "name": "GNU General Public License v3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/bingoBoard": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Create a new bingo board with random bingo cards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bingo Board"
                ],
                "summary": "Create bingo board.",
                "parameters": [
                    {
                        "description": "Bingo Board Name",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBingoBoardDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserDto"
                        }
                    }
                }
            }
        },
        "/bingoBoard/{id}": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Get a bingo board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bingo Board"
                ],
                "summary": "Get bingo board.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bingo Board ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Delete a bingo board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bingo Board"
                ],
                "summary": "Delete bingo board.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bingo Board ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserDto"
                        }
                    }
                }
            }
        },
        "/bingoBoard/{id}/addBingoCard": {
            "put": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add a bingo card to a bingo board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bingo Board"
                ],
                "summary": "Add bingo card.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bingo Board ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bingo Card UUID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BingoCardId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bingoBoard/{id}/join": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Join a bingo board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bingo Board"
                ],
                "summary": "Join bingo board.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bingo Board ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bingoBoard/{id}/leave": {
            "delete": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Leave a bingo board.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bingo Board"
                ],
                "summary": "Leave bingo board.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bingo Board ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/bingoCards": {
            "get": {
                "description": "Get all bingo cards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bingo Card"
                ],
                "summary": "Get all bingo cards.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BingoCardDto"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a bingo card.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bingo Card"
                ],
                "summary": "Create a bingo card.",
                "parameters": [
                    {
                        "description": "Bingo Card",
                        "name": "bingoCard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateBingoCardDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BingoCardDto"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Get information about the user that is currently logged in.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user information.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserDto"
                        }
                    }
                }
            }
        },
        "/me/bingoCard/{id}": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add or remove bingo card from user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Click bingo card.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bingo Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserDto"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.CreateBingoCardDto": {
            "type": "object",
            "required": [
                "description"
            ],
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "models.BingoBoardDto": {
            "type": "object",
            "properties": {
                "bingo_cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BingoCardDto"
                    }
                },
                "name": {
                    "type": "string"
                },
                "short_uuid": {
                    "type": "string"
                }
            }
        },
        "models.BingoCardDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "selected": {
                    "type": "boolean"
                },
                "user_count": {
                    "type": "integer"
                }
            }
        },
        "models.BingoCardId": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "models.CreateBingoBoardDto": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.UserDto": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "bingo_boards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BingoBoardDto"
                    }
                },
                "bingo_cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BingoCardDto"
                    }
                },
                "github_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reddit_url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}